Current state of visualization

+++ src/de/prochnow/instaScraper/VisualObject.java
@@ -3,9 +3,9 @@ package de.prochnow.instaScraper;
 
 public class VisualObject {
 
-    public int x = 0;
+    public float x = 0;
 
-    public int y = 0;
+    public float y = 0;
 
     public int color;
